server/requirements.txt
django==1.6
djangorestframework
django-cors-headers
django-extensions
ipython